feat: update cache for all property files if they point to the same changelog when contexts are loaded (#74)
* feat: update cache for all property files if they point to the same changelog when contexts are loaded

* chore: updated CHANGELOG.md and version

* chore: fix issues from PR

* feat: add selected contexts back, when contexts are loaded

* fix: add full path to find the real path in key

* test: fix open cache test

* chore: fix in/output error in tests

+++ src/test/e2e/LiquibaseGUITestUtils.ts
@@ -74,6 +74,8 @@ export class LiquibaseGUITestUtils {
       await DockerTestUtils.startContainer();
     }
 
+    this.removeContentOfFolder(path.join(this.WORKSPACE_PATH, "data", "liquibase"));
+
     // open the workspace
     await this.openWorkspace();
 
@@ -288,10 +290,13 @@ export class LiquibaseGUITestUtils {
    * @param folder - the folder
    */
   static removeContentOfFolder(folder: string): void {
-    for (const file of fs.readdirSync(folder)) {
-      fs.rmSync(path.join(folder, file), { recursive: true, force: true });
+    if (fs.existsSync(folder)) {
+      for (const file of fs.readdirSync(folder)) {
+        fs.rmSync(path.join(folder, file), { recursive: true, force: true });
+      }
     }
   }
+
   //#endregion
 
   //#region command end
